prevent underflow

+++ filetree.go
@@ -210,9 +210,9 @@ func (tree *FileTree) MarkRemoved(path string) error {
 	return node.AssignDiffType(Removed)
 }
 
-func StackRange(trees []*FileTree, index int) *FileTree {
+func StackRange(trees []*FileTree, index uint) *FileTree {
 	tree := trees[0].Copy()
-	for idx := 1; idx < index; idx++ {
+	for idx := uint(1); idx < index; idx++ {
 		tree.Stack(trees[idx])
 	}
 	return tree
